+++ packages/components/bitcoin/src/cqrs-components/events/network-cleared.event.ts
@@ -1,0 +1,6 @@
+import { SystemEvent, BasicEvent, EventBasePayload } from '@easylayer/common/cqrs';
+
+interface BitcoinNetworkClearedEventPayload extends EventBasePayload {}
+
+@SystemEvent()
+export class BitcoinNetworkClearedEvent extends BasicEvent<BitcoinNetworkClearedEventPayload> {}