+++ packages/components/evm/src/cqrs-components/events/index.ts
@@ -1,3 +1,4 @@
 export * from './network-blocks-added.event';
 export * from './network-initialized.event';
 export * from './network-reorganized.event';
+export * from './network-cleared.event';